+++ buildSrc/src/main/kotlin/Deps.kt
@@ -26,10 +26,10 @@ object Deps {
     private const val coroutinesVersion = "1.3.9"
     private const val mokoGraphicsVersion = "0.4.0"
     private const val mokoParcelizeVersion = "0.4.0"
-    private const val mokoResourcesVersion = "0.13.0"
+    private const val mokoResourcesVersion = "0.13.1"
     private const val mokoMvvmVersion = "0.8.0"
     private const val mokoFieldsVersion = "0.5.0"
-    private const val mokoUnitsVersion = "0.4.0"
+    private const val mokoUnitsVersion = "0.4.1"
     private const val mokoMediaVersion = "0.5.0"
     private const val mokoPermissionsVersion = "0.6.0"
     const val mokoWidgetsVersion = BuildConfig.widgetsVersion
