+++ moko-widgets-datetime-picker.podspec
@@ -11,7 +11,7 @@ Pod::Spec.new do |spec|
     spec.source_files             = "widgets-datetime-picker/src/iosMain/swift/**/*.{h,m,swift}"
     spec.resources                = "widgets-datetime-picker/src/iosMain/bundle/**/*"
 
-    spec.dependency 'FloatingPanel'
+    spec.dependency 'FloatingPanel', '~> 1.7.2'
 
     spec.ios.deployment_target  = '11.0'
     spec.swift_version          = '5.0'
