+++ widgets-bottomsheet/src/iosMain/kotlin/dev/icerock/moko/widgets/bottomsheet/BottomSheetHolder.kt
@@ -4,13 +4,26 @@
 
 package dev.icerock.moko.widgets.bottomsheet
 
+import cocoapods.mokoWidgetsBottomSheet.BottomSheetController
 import dev.icerock.moko.widgets.core.View
 import platform.UIKit.UIViewController
 
-internal expect class BottomSheetHolder() : SelfDismisser {
+internal class BottomSheetHolder : SelfDismisser {
+    private val bottomSheet = BottomSheetController()
+
     fun show(
         viewController: UIViewController,
         view: View,
         onDismiss: (isSelfDismissed: Boolean) -> Unit
-    )
+    ) {
+        bottomSheet.showOnViewController(
+            vc = viewController,
+            withContent = view,
+            onDismiss = onDismiss
+        )
+    }
+
+    override fun dismissSelf() {
+        bottomSheet.dismiss()
+    }
 }
