+++ widgets-collection/src/iosArm64Main
@@ -1,1 +1,1 @@
-iosX64Main
+iosMain
