--- widgets-collection/src/iosX64Main/kotlin/dev/icerock/moko/widgets/collection/ALCollectionFlowLayout.kt
@@ -1,11 +1,0 @@
-/*
- * Copyright 2020 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
- */
-
-package dev.icerock.moko.widgets.collection
-
-import platform.UIKit.UICollectionViewFlowLayout
-
-//import cocoapods.mokoWidgetsCollection.ALCollectionFlowLayout
-
-//actual typealias ALCollectionFlowLayout = UICollectionViewFlowLayout//ALCollectionFlowLayout
